# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒体同步和刮削系统API",
        "title": "diy-strm API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8115",
    "basePath": "/",
    "paths": {
...
        "/account/create": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的开放平台账号（支持115、OpenList等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号管理"
                ],
                "summary": "创建账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号源类型",
                        "name": "source_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "账号名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用ID（自定义时必需）",
                        "name": "app_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "应用ID名称（Q115-STRM、MQ的媒体库、自定义）",
                        "name": "app_id_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/account/delete": {
            "delete": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定ID的开放平台账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号管理"
                ],
                "summary": "删除账号",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/account/list": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有已配置的开放平台账号（如115、OpenList等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号管理"
                ],
                "summary": "获取账号列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/account/openlist": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的OpenList账号或更新现有账号的凭证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号管理"
                ],
                "summary": "创建/更新OpenList账号",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID（指定则为更新操作）",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "OpenList服务器地址",
                        "name": "base_url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}